feat: add map gender function to utils.py
