+++ src/scribe_data/extract_transform/languages/German/nouns/format_nouns.py
@@ -9,6 +9,11 @@ import collections
 import json
 import os
 import sys
+PATH_TO_SCRIBE_ORG = os.path.dirname(sys.path[0]).split("Scribe-Data")[0]
+PATH_TO_SCRIBE_DATA_SRC = f"{PATH_TO_SCRIBE_ORG}Scribe-Data/src"
+sys.path.insert(0, PATH_TO_SCRIBE_DATA_SRC)
+
+from scribe_data.utils import map_genders
 
 LANGUAGE = "German"
 QUERIED_DATA_TYPE = "nouns"
@@ -31,25 +36,6 @@ else:
 
 with open(data_path, encoding="utf-8") as f:
     nouns_list = json.load(f)
-
-
-def map_genders(wikidata_gender):
-    """
-    Maps those genders from Wikidata to succinct versions.
-
-    Parameters
-    ----------
-        wikidata_gender : str
-            The gender of the noun that was queried from WikiData
-    """
-    if wikidata_gender in ["masculine", "Q499327"]:
-        return "M"
-    elif wikidata_gender in ["feminine", "Q1775415"]:
-        return "F"
-    elif wikidata_gender in ["neuter", "Q1775461"]:
-        return "N"
-    else:
-        return ""  # nouns could have a gender that is not valid as an attribute
 
 
 def order_annotations(annotation):
